+++ src/effects/spells/darkness.js
@@ -3,12 +3,12 @@ import { loadMacroFile, generateMacroChange, generateItemMacroFlag } from "../ma
 
 export async function darknessEffect(document) {
   let effect = baseSpellEffect(document, document.name);
+  setProperty(effect, "flags.dae.selfTarget", true);
+  setProperty(effect, "flags.dae.selfTargetAlways", true);
   const itemMacroText = await loadMacroFile("spell", "darkness.js");
   document.flags["itemacro"] = generateItemMacroFlag(document, itemMacroText);
   effect.changes.push(generateMacroChange(""));
   document.effects.push(effect);
-  document.system['target']['type'] = "self";
-  document.system.range = { value: null, units: "self", long: null };
-
+  setProperty(document, "flags.midi-qol.onUseMacroName", "[preTargeting]ItemMacro");
   return document;
 }